fix: resolved a couple issues
